rename, refactor: Main.jsx 를 FolderPage.jsx 로 분리
