+++ src/components/FolderPage/Hero.jsx
