+++ src/components/FolderPage/SearchBar.css
