+++ src/components/FolderPage/SearchBar.jsx
