+++ src/components/FolderPage/index.js
@@ -1,0 +1,4 @@
+import Cards from "./Cards";
+import Hero from "./Hero";
+
+export { Cards, Hero };
