--- src/components/Main/index.js
@@ -1,3 +1,0 @@
-import Main from "./Main";
-
-export default Main;
